# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Komm — Bulk Messaging Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: groupId
          in: query
          schema:
            type: integer
        - name: channel
          in: query
          schema:
            type: string
            enum: [sms, whatsapp, email]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactsPage"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contacts/import:
    post:
      operationId: importContacts
      tags: [contacts]
      summary: Bulk import contacts from JSON (parsed from CSV/Excel client-side)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportContactsBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"

  # ── Groups ─────────────────────────────────────────────────────────────────
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List contact groups
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    put:
      operationId: updateGroup
      tags: [groups]
      summary: Update a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /groups/{id}/contacts:
    post:
      operationId: addContactsToGroup
      tags: [groups]
      summary: Add contacts to a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactIdsBody"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"

  # ── Templates ──────────────────────────────────────────────────────────────
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List message templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [sacco, church, general]
        - name: channel
          in: query
          schema:
            type: string
            enum: [sms, whatsapp, email]
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateBody"
      responses:
        "201":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateBody"
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Campaigns ──────────────────────────────────────────────────────────────
...
  /inbox:
    get:
      operationId: listInboxMessages
      tags: [inbox]
      summary: List incoming replies
      parameters:
        - name: read
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Inbox messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxPage"
  /inbox/{id}/read:
    post:
      operationId: markInboxRead
      tags: [inbox]
      summary: Mark a reply as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inbox message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxMessage"

  # ── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Contact:
      type: object
      required: [id, name, phone, channel, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        channel:
          type: string
          enum: [sms, whatsapp, email]
        groupIds:
          type: array
          items:
            type: integer
        customFields:
          type: object
          additionalProperties:
            type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateContactBody:
      type: object
      required: [name, phone, channel]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        channel:
          type: string
          enum: [sms, whatsapp, email]
        groupIds:
          type: array
          items:
            type: integer
        customFields:
          type: object
          additionalProperties:
            type: string
          nullable: true
    ContactsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ImportContactsBody:
      type: object
      required: [contacts]
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/CreateContactBody"
        groupId:
          type: integer
          nullable: true
    ImportResult:
      type: object
      required: [imported, duplicates, errors]
      properties:
        imported:
          type: integer
        duplicates:
          type: integer
        errors:
          type: integer
        errorDetails:
          type: array
          items:
            type: string
    ContactIdsBody:
      type: object
      required: [contactIds]
      properties:
        contactIds:
          type: array
          items:
            type: integer

    # ── Groups ──────────────────────────────────────────────────────────────
    Group:
      type: object
      required: [id, name, contactCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        contactCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateGroupBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true

    # ── Templates ────────────────────────────────────────────────────────────
    Template:
      type: object
      required: [id, name, body, channel, category, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        body:
          type: string
        channel:
          type: string
          enum: [sms, whatsapp, email]
        category:
          type: string
          enum: [sacco, church, general]
        variables:
          type: array
          items:
            type: string
          description: Variable placeholder names e.g. ["name","amount","date"]
        createdAt:
          type: string
          format: date-time
    CreateTemplateBody:
      type: object
      required: [name, body, channel, category]
      properties:
        name:
          type: string
        body:
          type: string
        channel:
          type: string
          enum: [sms, whatsapp, email]
        category:
          type: string
          enum: [sacco, church, general]
        variables:
          type: array
          items:
            type: string

    # ── Campaigns ────────────────────────────────────────────────────────────
...
    InboxMessage:
      type: object
      required: [id, from, body, channel, read, receivedAt]
      properties:
        id:
          type: integer
        from:
          type: string
        contactName:
          type: string
          nullable: true
        body:
          type: string
        channel:
          type: string
          enum: [sms, whatsapp]
        read:
          type: boolean
        campaignId:
          type: integer
          nullable: true
        receivedAt:
          type: string
          format: date-time
    InboxPage:
      type: object
      required: [data, total, page, limit, unreadCount]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InboxMessage"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        unreadCount:
          type: integer

    # ── Dashboard ────────────────────────────────────────────────────────────
...