# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Komm — Bulk Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ContactChannel } from "./contactChannel";
import type { ContactCustomFields } from "./contactCustomFields";

export interface Contact {
  id: number;
  name: string;
  phone: string;
  email?: string | null;
  channel: ContactChannel;
  groupIds?: number[];
  customFields?: ContactCustomFields;
  createdAt: Date;
}