# lib/api-zod/src/generated/types/createContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Komm — Bulk Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContactBodyChannel } from "./createContactBodyChannel";
import type { CreateContactBodyCustomFields } from "./createContactBodyCustomFields";

export interface CreateContactBody {
  name: string;
  phone: string;
  email?: string | null;
  channel: CreateContactBodyChannel;
  groupIds?: number[];
  customFields?: CreateContactBodyCustomFields;
}