# lib/api-zod/src/generated/types/createGroupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Komm — Bulk Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGroupBody {
  name: string;
  description?: string | null;
}